+++ podium-uaa/src/main/java/nl/thehyve/podium/web/rest/ElasticsearchIndexResource.java
@@ -17,7 +17,7 @@ package nl.thehyve.podium.web.rest;
 import com.codahale.metrics.annotation.Timed;
 import nl.thehyve.podium.common.security.AuthorityConstants;
 import nl.thehyve.podium.common.security.annotations.SecuredByAuthority;
-import nl.thehyve.podium.security.SecurityService;
+import nl.thehyve.podium.common.service.SecurityService;
 import nl.thehyve.podium.service.ElasticsearchIndexService;
 import nl.thehyve.podium.web.rest.util.HeaderUtil;
 import org.slf4j.Logger;
